Refactor: reorganize Folder Structure And Rename Scripts
