+++ Assets/Resources/Data/Dancepace/WaveData.asset
@@ -1,0 +1,57 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 43bb819376939bc40b85b51c15d25c22, type: 3}
+  m_Name: WaveData
+  m_EditorClassIdentifier: Scripts:Runtime.CH3.Dancepace:DPWaveDataSO
+  waveDatas:
+  - waveId: Wave_Normal_0
+    beats:
+    - poseData: 1
+      timing: 1
+      restTime: 0.5
+    - poseData: 1
+      timing: 0.5
+      restTime: 0.5
+    - poseData: 2
+      timing: 0.5
+      restTime: 0.5
+    - poseData: 3
+      timing: 0.5
+      restTime: 1
+  - waveId: Wave_Normal_1
+    beats:
+    - poseData: 1
+      timing: 1
+      restTime: 0.5
+    - poseData: 1
+      timing: 0.5
+      restTime: 0.5
+    - poseData: 2
+      timing: 0.5
+      restTime: 0.5
+    - poseData: 3
+      timing: 0.5
+      restTime: 1
+  - waveId: Wave_Normal_2
+    beats:
+    - poseData: 2
+      timing: 1
+      restTime: 1
+    - poseData: 4
+      timing: 0.5
+      restTime: 0.25
+    - poseData: 4
+      timing: 0.5
+      restTime: 1
+    - poseData: 1
+      timing: 1
+      restTime: 0.5
