+++ Assets/Resources/Data/SuperArio.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5ea7de979fe4534a9536571f5f5e77e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
