+++ Assets/Resources/Data/SuperArio/ObstacleSpawnDataSet.asset
@@ -1,0 +1,51 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d075088e075e27d46a22a83d0e6430db, type: 3}
+  m_Name: ObstacleSpawnDataSet
+  m_EditorClassIdentifier: 
+  DataList:
+  - Stage: 1-1
+    Speed: 1.5
+    ObstacleTypes: 3
+    ObstacleCount: 10
+  - Stage: 1-2
+    Speed: 2.5
+    ObstacleTypes: 3
+    ObstacleCount: 15
+  - Stage: 1-3
+    Speed: 3
+    ObstacleTypes: 3
+    ObstacleCount: 20
+  - Stage: 2-1
+    Speed: 3
+    ObstacleTypes: 6
+    ObstacleCount: 15
+  - Stage: 2-2
+    Speed: 3.5
+    ObstacleTypes: 6
+    ObstacleCount: 15
+  - Stage: 2-3
+    Speed: 4
+    ObstacleTypes: 6
+    ObstacleCount: 20
+  - Stage: 3-1
+    Speed: 4.5
+    ObstacleTypes: 6
+    ObstacleCount: 15
+  - Stage: 3-2
+    Speed: 5
+    ObstacleTypes: 6
+    ObstacleCount: 20
+  - Stage: 3-3
+    Speed: 5.5
+    ObstacleTypes: 6
+    ObstacleCount: 30
