--- Assets/Resources/SO/Dancepace/GameConfigData.asset
@@ -1,22 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b4b898bc9b63c864b89e6c04ccd3411b, type: 3}
-  m_Name: GameConfigData
-  m_EditorClassIdentifier: Scripts:Runtime.CH3.Dancepace:DPGameConfig
-  limitTime: 60
-  waveForCount: 6
-  waveClearCoin: 15
-  greatCoin: 3
-  goodCoin: 2
-  badCoin: 0
-  greatTimingWindow: 0.1
-  goodTimingWindow: 0.2
