--- Assets/Resources/SO/Dancepace/GameConfigData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: add2c5130c6fa6c439a597e0087e3b8b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
