--- Assets/Resources/SO/Dancepace/WaveData.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0e661170600e0b548aef0e40af8131f7
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
