--- Assets/Resources/SO/ObstacleSpawnDataSet.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 61e1e2197d247f249acd4c127df08455
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
