+++ Assets/Scenes/CH3.unity
@@ -420,7 +420,7 @@ GameObject:
   m_Layer: 0
   m_Name: Player
   m_TagString: Player
-  m_Icon: {fileID: 0}
+  m_Icon: {fileID: -5487077368411116049, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
@@ -504,7 +504,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: ca6dc948b88409142a943e233dbadc78, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  moveSpeed: 5
+  moveSpeed: 10
   virtualCamera: {fileID: 793652120}
 --- !u!65 &245858297
 BoxCollider:
@@ -1118,7 +1118,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   minimapCamera: {fileID: 1516391664}
-  minimapSize: 10
   minimapLayers:
     serializedVersion: 2
     m_Bits: 262144
@@ -1482,7 +1481,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 716496751}
   serializedVersion: 2
-  m_LocalRotation: {x: 0.18687533, y: 0, z: 0, w: 0.98238367}
+  m_LocalRotation: {x: 0.18687533, y: 1e-45, z: 0, w: 0.98238367}
   m_LocalPosition: {x: 0, y: 4.07, z: -8.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -1631,7 +1630,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 793652119}
   serializedVersion: 2
-  m_LocalRotation: {x: 0.18687533, y: -0, z: -0, w: 0.98238367}
+  m_LocalRotation: {x: 0.18687533, y: 1e-45, z: -0, w: 0.98238367}
   m_LocalPosition: {x: 0, y: 3.75, z: -9.5}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -2221,7 +2220,7 @@ GameObject:
   m_Layer: 0
   m_Name: GridAreaSpawner
   m_TagString: Untagged
-  m_Icon: {fileID: 0}
+  m_Icon: {fileID: 7174288486110832750, guid: 0000000000000000d000000000000000, type: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
@@ -2284,7 +2283,6 @@ MonoBehaviour:
   cellWidth: 1
   heightOffset: 0.32
   mainCamera: {fileID: 716496753}
-  showDebugVisuals: 1
   gridLineColor: {r: 0.5, g: 0.5, b: 0.5, a: 0.5}
 --- !u!1 &1214100539
 GameObject:
