+++ Assets/Scripts/Runtime/CH3/Dancepace/Characters/DPRapley.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using UnityEngine.InputSystem;
+using Runtime.CH3.Dancepace;
 
 namespace Runtime.CH3.Dancepace
 {
