--- Assets/Scripts/Runtime/CH3/Dancepace/DPOpening.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4db9b365679ab17459020778f8c39e5a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
