--- Assets/Scripts/Runtime/CH3/Dancepace/Data/DPGameConfigSO.cs
@@ -1,17 +1,0 @@
-using UnityEngine;
-
-namespace Runtime.CH3.Dancepace
-{
-    [CreateAssetMenu(menuName = "Dancepace/GameConfig")]
-    public class DPGameConfigSO : ScriptableObject
-    {
-        public float limitTime;
-        public int waveForCount;
-        public int waveClearCoin;
-        public int greatCoin;
-        public int goodCoin;
-        public int badCoin;
-        public float greatTimingWindow;
-        public float goodTimingWindow;
-    }
-}
