--- Assets/Scripts/Runtime/CH3/Dancepace/Data/DPGameConfigSO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b4b898bc9b63c864b89e6c04ccd3411b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
