+++ Assets/Scripts/Runtime/CH3/Dancepace/Data/WaveDataSO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 43bb819376939bc40b85b51c15d25c22
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
