+++ Assets/Scripts/Runtime/CH3/Dancepace/Effects/EffectController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e11cebbc57b38a04f9bff4e7bda813c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
