+++ Assets/Scripts/Runtime/CH3/Dancepace/Effects/HeartEffect.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using DG.Tweening;
+using Runtime.CH3.Dancepace;
 
 namespace Runtime.CH3.Dancepace
 {
