--- Assets/Scripts/Runtime/CH3/Dancepace/Managers/DPManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c96a6f1259385ec4893fa03a1a50b357
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
