--- Assets/Scripts/Runtime/CH3/Dancepace/Managers/DPUIManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b7e2278bce21a844ac05f0f14aba0c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
