--- Assets/Scripts/Runtime/CH3/Dancepace/SpeakerAnimation.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c8e8a0feb3038144688e1c91744e81e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
